tsdb/querier: reuse intervals slice
```go
	p.bufIter.Intervals = p.bufIter.Intervals[:0]
	for _, interval := range p.intervals {
		if p.currMeta.OverlapsClosedInterval(interval.Mint, interval.Maxt) {
			p.bufIter.Intervals = p.bufIter.Intervals.Add(interval, true)
		}
	}
```

The querier code tries to reuse `p.bufIter.Intervals` but `Add()` before
this change always allocated a new slice. I added a parameter which
controls whether the same slice is reused.

I noticed this through profiling - even though we don't use tombstones,
this function in this was place was allocating ~6% of all objects.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>